--- pages/posts.js
@@ -1,29 +1,0 @@
-/* eslint-disable react-hooks/exhaustive-deps */
-import React, { useState, useEffect } from 'react';
-import { getPosts } from '../api/fbPostData';
-// import { useAuth } from '../utils/context/authContext';
-import PostCard from '../components/PostCard';
-
-export default function Posts() {
-  const [posts, setPosts] = useState([]);
-  // const { user } = useAuth();
-
-  const getThePosts = () => {
-    getPosts().then(setPosts);
-  };
-
-  useEffect(() => {
-    getThePosts();
-  }, []);
-
-  return (
-    <div>
-      <div className="text-center my-4">
-        <h1> POSTS </h1>
-        <div className="d-flex flex-wrap">
-          {posts.map((post) => <PostCard postObj={post} key={post.firebaseKey} onUpdate={getThePosts} />)}
-        </div>
-      </div>
-    </div>
-  );
-}
